scripts/requirements.txt
# AITMPL Manager Dependencies
requests>=2.28.0
pyyaml>=6.0
